enable M470 emulation for M470BSD/BSE/BSS variants; simplify configuration
This adds a CONFIG_MACH_M470_EMULATION option to permit all M470 variants to
use the same kernel and Android ROM (provided they use identical hardware).
When enabled, variants will report their machine names as 'm470' rather than
'm470xxx'. To work, both MACH_M470 and all desired variants must be enabled.

This patch also simplifies machine-dependent code by replacing Hisense's
jumble of multiple #ifdefs that used MACH_M470XXX and BOARD_M470XXX inter-
changably with a simple #ifdef BOARD_M470 that applies to all variants.

Note that this methodology could not be applied to the machine_is_m470xxx()
macros; instead, additional code had to be added to support each variant.
In fact, it was because of the way these macros are generated and resolved
by the build process that the seemingly simpler method of replacing each
variant's machine ID with that of the M470 was not used.