exfat: enable module to support vfat (fat12/16/32)
Note: the changes to 'm470_v2_dopa1_defconfig' which enable this support
to be built are in commit bb39f0c89363761cc3b09d3870ab10a0fcd18cd8
("enable exFAT module to support vfat" dated 29 Sep 2014). Because of
the many intervening changes to the defconfig, that patch could not be
reverted and merged with this one, as I would have preferred.