build: fix 'maybe-uninitialized' warnings
This fixes warnings generated by linaro 4.9 and gcc 5.1, primarily
when optimization is set to -O3. Most of these can be suppressed
without generating any code using the 'uninitialized_var(x)' macro
found in compiler-gcc.h (so it claims). It is effective for scalars
(e.g. ints or ptrs) but not for compound variables (i.e. structs);
for those, the affected member has to be explicitly initialized.