vfs: mnt mega-patch
This patch was created by replacing all affected files in the 'fs'
directory with versions found in the Linux source as of the end of
Jan 04, 2012. Interdependent files in other locations were then
replaced as needed. The goal was to avoid applying over 100 patches,
many of which were irrelevant to the ultimate goal of implementing
mount namespaces.